requirements.txt
pyTelegramBotAPI
google-generativeai>=0.8.0

